pyproject.toml
[project]
name = "driving-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-webhook>=1.4.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
